ref: rename `ifg` to `calculator`
