+++ ifg_py/__init__.py
@@ -1,4 +1,4 @@
-from ifg_py.ifg import IfgCalculator
+from ifg_py.calculator import IfgCalculator
 from ifg_py.units_converter import SiAtomicConverter, get_metal_specific_volume
 
 from ifg_py.metadata import __version__
