+++ tests/unit/test_around_threshold.py
@@ -1,4 +1,4 @@
-from ifg_py.ifg import THRESHOLD
+from ifg_py.calculator import THRESHOLD
 from hypothesis import strategies as st, given
 from tests.conftest import volumes_st, set_up
 import numpy as np
